Move index_search to _core
