+++ parcels/_core/xgrid.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 import xarray as xr
 import xgcm
 
-from parcels._index_search import _search_indices_curvilinear_2d
+from parcels._core.index_search import _search_indices_curvilinear_2d
 from parcels._typing import assert_valid_mesh
 from parcels._core.basegrid import BaseGrid
 from parcels._core.utils.array import _search_1d_array
