+++ tests/test_index_search.py
@@ -4,7 +4,7 @@ import xarray as xr
 import xgcm
 
 from parcels._datasets.structured.generic import datasets
-from parcels._index_search import _search_indices_curvilinear_2d
+from parcels._core.index_search import _search_indices_curvilinear_2d
 from parcels._tutorial import download_example_dataset
 from parcels import Field
 from parcels import XGrid
